#37569709 Скорректировал логику (но остались некорректности при распределении). Также ускорил загрузку раб.пространства.

+++ LeokaEstetica.Platform.Models/Dto/Base/ProjectManagement/BaseTaskStatusOutput.cs
@@ -8,7 +8,7 @@ public class BaseTaskStatusOutput
     /// <summary>
     /// Id статуса.
     /// </summary>
-    public long StatusId { get; set; }
+    public int StatusId { get; set; }
     
     /// <summary>
     /// Id статуса, который маппится на статус задачи.
